Move generateMermaidGraph
