+++ lib/modules/manager/pip-compile/common.ts
@@ -7,11 +7,7 @@ import { ensureCacheDir } from '../../../util/fs';
 import * as hostRules from '../../../util/host-rules';
 import { regEx } from '../../../util/regex';
 import type { PackageFileContent, UpdateArtifactsConfig } from '../types';
-import type {
-  DependencyBetweenFiles,
-  GetRegistryUrlVarsResult,
-  PipCompileArgs,
-} from './types';
+import type { GetRegistryUrlVarsResult, PipCompileArgs } from './types';
 
 export function getPythonConstraint(
   config: UpdateArtifactsConfig,
@@ -215,24 +211,6 @@ function throwForUnknownOption(arg: string): void {
     return;
   }
   throw new Error(`Option ${arg} not supported (yet)`);
-}
-
-export function generateMermaidGraph(
-  depsBetweenFiles: DependencyBetweenFiles[],
-  lockFileArgs: Map<string, PipCompileArgs>,
-): string {
-  const lockFiles = [];
-  for (const lockFile of lockFileArgs.keys()) {
-    // TODO: add extra args to the lock file ${extraArgs ? '\n' + extraArgs : ''}
-    // const extraArgs = pipCompileArgs.extra
-    //   ?.map((v) => '--extra=' + v)
-    //   .join('\n');
-    lockFiles.push(`  ${lockFile}[[${lockFile}]]`);
-  }
-  const edges = depsBetweenFiles.map(({ sourceFile, outputFile, type }) => {
-    return `  ${sourceFile} -${type === 'constraint' ? '.' : ''}-> ${outputFile}`;
-  });
-  return `graph TD\n${lockFiles.join('\n')}\n${edges.join('\n')}`;
 }
 
 function buildRegistryUrl(url: string): URL | null {
